refactor: move capability builder to core
